+++ GoGreen/Controllers/GreenIslandController.cs
@@ -70,19 +70,12 @@ namespace GoGreen.Controllers
         }
 
         // POST: api/GreenIsland
-        
+        [Authorize]
         [HttpPost]
         //[Consumes("multipart/form-data")]
         public async Task<ActionResult<GreenIslandResponse>> Post([FromBody] GreenIslandRequest request)
         {
             //, IFormFile imageFile
-
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
-            if (string.IsNullOrEmpty(userId))
-            {
-                return BadRequest("The user ID claim is missing");
-            }
 
             var data = _mapper.Map<GreenIslandRequest>(request);
 
@@ -115,21 +108,14 @@ namespace GoGreen.Controllers
         }
 
         // PUT: api/GreenIsland/5
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult<GreenIslandResponse>> Put(int id, [FromBody] GreenIslandRequest request)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(id))
             {
-                return BadRequest("The user ID claim is missing");
-            }
-            var MunicipalityId = 2; //TODO
-            var municipality = await _context.Municipalities.FindAsync(MunicipalityId);
-
-            if (municipality == null)
-            {
-                return BadRequest($"The Municipality with ID {MunicipalityId} does not exist");
+                return BadRequest("The user has no permission for this action");
             }
 
             var updatedData = await _greenIslandService.Update(id, request);
@@ -139,9 +125,15 @@ namespace GoGreen.Controllers
         }
 
         // DELETE: api/GreenIsland/5
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+
+            if (!await CheckPermisionAsync(id))
+            {
+                return BadRequest("The user has no permission for this action");
+            }
 
             var isDeleted = await _greenIslandService.Delete(id);
 
@@ -153,20 +145,20 @@ namespace GoGreen.Controllers
             return NoContent();
         }
 
+        [Authorize]
         [HttpPut("{greenIslandId}/Image")]
         [Consumes("multipart/form-data")]
         public async Task<ActionResult<GreenIslandResponse>> AddImage(int greenIslandId, IFormFile imageFile)
 
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(greenIslandId))
             {
-                return BadRequest("The user ID claim is missing");
+                return BadRequest("The user has no permission for this action");
             }
 
             var data = await _context.GreenIslands
-                .Where(e => e.Id == greenIslandId && e.UserId == userId)
+                .Where(e => e.Id == greenIslandId)
                 .Include(a => a.GreenIslandImages)
                     .ThenInclude(ei => ei.Image)
                 .FirstOrDefaultAsync();
@@ -204,19 +196,18 @@ namespace GoGreen.Controllers
 
 
         // DELETE: api/GreenIsland/5/Image/2
+        [Authorize]
         [HttpDelete("{greenIslandId}/Image/{imageId}")]
         public async Task<IActionResult> DeleteEventImage(int greenIslandId, int imageId)
         {
 
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(greenIslandId))
             {
-                return BadRequest("The user ID claim is missing");
+                return BadRequest("The user has no permission for this action");
             }
 
             var data = await _context.GreenIslands
-                .Where(e => e.Id == greenIslandId && e.UserId == userId)
+                .Where(e => e.Id == greenIslandId)
                 .Include(e => e.GreenIslandImages)
                     .ThenInclude(ei => ei.Image)
                 .FirstOrDefaultAsync();
@@ -242,5 +233,29 @@ namespace GoGreen.Controllers
             return NoContent();
         }
 
+        private async Task<bool> CheckPermisionAsync(int id)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+
+            if (string.IsNullOrEmpty(userId))
+            {
+                return false;
+            }
+
+            var user = await _context.User.Include(e => e.Municipality).FirstOrDefaultAsync(u => u.Id == userId);
+            var data = await _context.GreenIslands
+                        .Where(a => a.MunicipalityId == user.MunicipalityId)
+                        .Where(a => a.Id == id)
+                        .FirstOrDefaultAsync();
+
+            if (data == null)
+            {
+                return false;
+            }
+
+            return true;
+
+        }
     }
 }
